requirements.txt
albumentations>=1.4.0
numpy>=1.26.0
opencv-python-headless>=4.10.0
pandas>=2.2.0
Pillow>=10.0.0
pyyaml>=6.0.0
scikit-learn>=1.5.0
scipy>=1.13.0
timm>=1.0.0
torch>=2.3.0
torchvision>=0.18.0
tqdm>=4.66.0
transformers>=4.44.0
